+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Add `application/x-www-form-urlencoded` body type handling, by @HardNorth
 
 ## [5.0.0]
 ### Added